requirements.txt
esptool==2.6
PyQt5==5.11.3
PyQt5-sip==4.19.15
pyqt5-tools==5.11.3.1.4
pyserial==3.4
